+++ packages/core-data/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 2.1.0 (Unreleased)
 
 ### New features
-- The `getAutosave` and `hasAutosave` selectors were added.
+- The `getAutosave` selector has been added.
 - The `receiveAutosave` action has been added.
 
 ## 2.0.16 (2019-01-03)