+++ packages/editor/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Deprecations
 
-- The `getAutosave`, `getAutosaveAttribute`, and `hasAutosave` selectors are deprecated. Equivalent selectors have been added to the `@wordpress/core-data` package.
+- The `getAutosave`, `getAutosaveAttribute`, and `hasAutosave` selectors are deprecated. Please use the `getAutosave` selector in the `@wordpress/core-data` package.
 - The `resetAutosave` action is deprecated. An equivalent action `receiveAutosave` has been added to the `@wordpress/core-data` package.
 - The `isEditedPostAutosaveable` action now requires that the parameter `autosave` is provided.